+++ packages/ngx-icon/src/lib/icon.module.ts
@@ -1,5 +1,5 @@
 import { NgModule } from '@angular/core';
-import {CommonModule} from '@angular/common';
+import { CommonModule } from '@angular/common';
 import { IconComponent } from './icon.component';